% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{arrayDataDEA_limma}
\alias{arrayDataDEA_limma}
\title{arrayDataDEA_limma}
\usage{
arrayDataDEA_limma(data, group, comparison, cutFC = 2, cutFDR = 0.05)
}
\arguments{
\item{data}{matrix}

\item{group}{a vector}

\item{comparison}{character}

\item{cutFC}{>1}

\item{cutFDR}{0< cutFDR <0.05}
}
\value{
data.frame
}
\description{
arrayDataDEA_limma
}
