% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFUNscore.R
\name{calFUNscore}
\alias{calFUNscore}
\title{Calculate Functional score using scRNA or bulk transcriptomic data}
\usage{
calFUNscore(
  expr,
  fungeneset = NULL,
  study.type = NULL,
  num.rounds = 1000,
  n.bins = 50,
  b.sign.num = 1,
  seed = TRUE
)
}
\arguments{
\item{expr}{TPM matrix derived from scRNAseq, or normalized gene expression matrix from bulk samples; rows are genes, cols are cells or samples.}

\item{fungeneset}{The character vectors containing selected function-related genes.}

\item{study.type}{A string value to indicate the study type for calculation. Allowed values contain c('scRNAseq', 'bulk_RNAseq').}

\item{num.rounds}{A integer value to indicate the permutation times of iteration; 1000 by default and 10000 will be better for reproducibility.}

\item{n.bins}{A integer value giving expression bins within genes, which were binned according to their average expression across cells or samples. Set to 50 by default.}

\item{seed}{A logical value to indicate if set random seed for reproducible results. TRUE by default.}
}
\value{
Continuous numerical variables with FUNscore, named by cells or samples.
}
\description{
Calculate Functional score using scRNA or bulk transcriptomic data
}
\examples{
\dontrun{
data("expr")
data("funRelatedGeneSet")
fungeneset = funRelatedGeneSet[["Angiogenesis"]]@geneIds
cfs <- calFUNscore(expr = expr, fungeneset = fungeneset, study.type = "bulk_RNAseq")
}
}
\author{
Leihe, Yixian Fan,Sen Chen.
}
