% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{Counts2TPM}
\alias{Counts2TPM}
\title{Counts2TPM}
\usage{
Counts2TPM(counts, effLen)
}
\arguments{
\item{counts}{a data.frame or matrix for raw count of RNAseq.}

\item{effLen}{The length of genes.}
}
\value{
a data.frame
}
\description{
Counts2TPM
}
