% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{cox.forestplot}
\alias{cox.forestplot}
\title{cox.forestplot}
\usage{
cox.forestplot(forest_data, ylab, save = FALSE, folder = ".")
}
\arguments{
\item{forest_data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{ylab}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
cox.forestplot
}
