% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{delTCGA_dup_sample}
\alias{delTCGA_dup_sample}
\title{In the expression matrix of TCGA tumor samples, the data of duplicate patients were deleted.}
\usage{
delTCGA_dup_sample(data, col_rename = T)
}
\arguments{
\item{data}{Gene expression matrix of TCGA tumor samples.}

\item{col_rename}{TRUE or FALSE}
}
\value{
data.frame
}
\description{
In the expression matrix of TCGA tumor samples, the data of duplicate patients were deleted.
}
