% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{dlTCGAmiRNAdata}
\alias{dlTCGAmiRNAdata}
\title{dlTCGAmiRNAdata}
\usage{
dlTCGAmiRNAdata(data)
}
\arguments{
\item{data}{data is a data frame where the first column represents the names of miRNAs, the second column contains the corresponding values, and the third column is the barcode.}
}
\description{
dlTCGAmiRNAdata
}
