% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichAnalysis.R
\name{exe.gseGO_GSEA}
\alias{exe.gseGO_GSEA}
\title{exe.gseGO_GSEA}
\usage{
exe.gseGO_GSEA(
  data,
  gseGO.ont = "BP",
  keyType = "SYMBOL",
  OrgDb = org.Hs.eg.db,
  species = "Homo sapiens",
  TERM2GENE = "msigdbr",
  category = "C5",
  subcategory = "GO:BP",
  showCategory = 6,
  pvalueCutoff = 0.01,
  fileName = "enrich",
  save = TRUE,
  height = 4,
  width = 7,
  folder = "./"
)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{gseGO.ont}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{keyType}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{OrgDb}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{species}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TERM2GENE}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{category}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{subcategory}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{showCategory}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{pvalueCutoff}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{fileName}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{height}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{width}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
exe.gseGO_GSEA
}
