% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichAnalysis.R
\name{exeGO_KEGG}
\alias{exeGO_KEGG}
\title{exeGO_KEGG}
\usage{
exeGO_KEGG(
  geneset,
  geneType = "SYMBOL",
  organism = "hsa",
  showCategory = 6,
  pvalueCutoff = 0.05,
  OrgDb = "org.Hs.eg.db",
  KEGG = TRUE,
  save = TRUE,
  folder = "./",
  Prefix = "enrich",
  height = 6,
  width = 10
)
}
\arguments{
\item{geneset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{geneType}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{organism}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{showCategory}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{pvalueCutoff}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{OrgDb}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{KEGG}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{Prefix}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{height}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{width}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
exeGO_KEGG
}
