% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichAnalysis.R
\name{exeGO_KEGG2}
\alias{exeGO_KEGG2}
\title{exeGO_KEGG2}
\usage{
exeGO_KEGG2(
  geneset,
  keyType = "SYMBOL",
  organism = "hsa",
  showCategory = 4,
  color = NULL,
  pvalueCutoff = 0.05,
  OrgDb = "org.Hs.eg.db",
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  KEGG = TRUE,
  save = TRUE,
  folder = "./",
  Prefix = "enrich",
  height = 5,
  width = 5.5
)
}
\arguments{
\item{geneset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{keyType}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{organism}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{showCategory}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{color}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{pvalueCutoff}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{OrgDb}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{pAdjustMethod}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{qvalueCutoff}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{minGSSize}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{maxGSSize}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{readable}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{KEGG}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{Prefix}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{height}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{width}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
exeGO_KEGG2
}
