% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{featureSelect.lasso}
\alias{featureSelect.lasso}
\title{featureSelect.lasso}
\usage{
featureSelect.lasso(
  data,
  dataFrom = "mergeSurExp",
  feature = "all",
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{dataFrom}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{feature}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
vector
}
\description{
featureSelect.lasso
}
