% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{filterGeneTypeExpr}
\alias{filterGeneTypeExpr}
\title{RNAseq data in TCGA database were filtered according to genotype.}
\usage{
filterGeneTypeExpr(expr, fil_col = "gene_type", filter = FALSE)
}
\arguments{
\item{expr}{Gene expression data.}

\item{fil_col}{A string. Column name to be filtered.}

\item{filter}{A value in the fil_col column.}
}
\description{
RNAseq data in TCGA database were filtered according to genotype.
}
