% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{geneDEAnalysis}
\alias{geneDEAnalysis}
\title{A function for differential expression analysis}
\usage{
geneDEAnalysis(data, group, comparison, method = "DESeq2", filter = TRUE)
}
\arguments{
\item{data}{for matrix input: a matrix of non-negative integers}

\item{group}{for matrix input: a DataFrame or data.frame with at least a single column. Rows of group correspond to columns of countData}

\item{comparison}{A string linked by "-" that represents the grouping information in the group. Such as "tumor-normal".}

\item{method}{One of DESeq2, edgeR, and limma.}

\item{filter}{TRUE or FALSE}
}
\value{
A data.frame
}
\description{
A function for differential expression analysis
}
