% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{geneset2gmt}
\alias{geneset2gmt}
\title{geneset2gmt}
\usage{
geneset2gmt(
  geneset,
  genesetname,
  description = NA,
  return = "data.frame",
  folder = ".",
  filename
)
}
\arguments{
\item{geneset}{A vector of gene names or a path to a txt file (file content should contain one gene per line)}

\item{genesetname}{A character string specifying the name of the gene set}

\item{description}{A character string describing the gene set, default is NA}

\item{return}{The type of object to return, either "data.frame" or "GeneSetCollection"}

\item{folder}{Path to the output folder, default is current working directory}

\item{filename}{Name of the output GMT file (must end with .gmt)}
}
\value{
An object of the type specified by the 'return' parameter
}
\description{
geneset2gmt
}
