% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{GeneSetCollection.to.df}
\alias{GeneSetCollection.to.df}
\title{GeneSetCollection.to.df}
\usage{
GeneSetCollection.to.df(GeneSetCollection)
}
\arguments{
\item{GeneSetCollection}{GeneSetCollection object.}
}
\value{
data.frame
}
\description{
GeneSetCollection.to.df
}
