% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{get.geneExpbasedonCor}
\alias{get.geneExpbasedonCor}
\title{get.geneExpbasedonCorget.geneExpbasedonCor}
\usage{
get.geneExpbasedonCor(
  expr,
  gene,
  TCGA = FALSE,
  filter = "protein_coding",
  cor.method = "pearson",
  cor.coef = 0.3,
  cor.p.value = 0.05,
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{gene}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filter}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.coef}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.p.value}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{exprFor}{more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
get.geneExpbasedonCorget.geneExpbasedonCor
}
