% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_dataDownload.R
\name{get_miRNAEQ}
\alias{get_miRNAEQ}
\title{get_miRNAEQ}
\usage{
get_miRNAEQ(project, save = FALSE, folder = ".")
}
\arguments{
\item{project}{A list of valid project (see list with TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{save}{TRUE or FALSE}

\item{folder}{The path specifies the location of the folder where the data is saved when the 'save' parameter is set to TRUE.}
}
\description{
get_miRNAEQ
}
