% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseAnalyses.R
\name{get_random_FUN}
\alias{get_random_FUN}
\title{Get random values from permutation test,  the mean of which as baseline to calculate FUN score.}
\usage{
get_random_FUN(input.list, genes.dist.bins, b.sign, num.rounds = 1000)
}
\arguments{
\item{input.list}{A list of variables containing scaled gene expression matrix, giving number of bins.}

\item{genes.dist.bins}{A matrix of giving expression bins within genes, which were binned according to their average expression across cells or samples.}

\item{b.sign}{A logical value to indicate the overlapped features between target features and the features derived from gene expressio matrix.}

\item{num.rounds}{A integer value to indicate the permutation times of iteration; 1000 by default and 10000 will be better for reproducibility.}
}
\value{
Continuous numerical variables with values, named by cells or samples.
}
\description{
Get random values from permutation test,  the mean of which as baseline to calculate FUN score.
}
\examples{

\dontrun{
r.scores <- get_random_FUN(input.list, input.list$genes.dist.bins, b.sign, num.rounds = num.rounds)
}

}
