% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_dataDownload.R
\name{getClinicalData}
\alias{getClinicalData}
\title{getClinicalData}
\usage{
getClinicalData(project, save = FALSE, folder = ".", trim = TRUE)
}
\arguments{
\item{project}{A list of valid project (see list with TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{save}{TRUE or FALSE}

\item{folder}{The path specifies the location of the folder where the data is saved when the 'save' parameter is set to TRUE.}

\item{trim}{TRUE or FALSE}
}
\description{
getClinicalData
}
