% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFileProcessing.R, R/baseFun.R
\name{getGeneBaseInfo}
\alias{getGeneBaseInfo}
\title{getGeneBaseInfo}
\usage{
getGeneBaseInfo(gtf)

getGeneBaseInfo(gtf)
}
\arguments{
\item{gtf}{The gtf parameter represents the path to a GTF (Gene Transfer Format) format file.}
}
\value{
a data.frame

data.frame
}
\description{
getGeneBaseInfo

getGeneBaseInfo
}
