% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAdataprepare.R
\name{getGeneExpData.pancancer}
\alias{getGeneExpData.pancancer}
\title{getGeneExpData.pancancer}
\usage{
getGeneExpData.pancancer(
  datafolder,
  geneSymbol,
  projects = "All",
  geneType = "protein_coding",
  dataType = "tpm",
  pattern = "STARdata.Rdata$",
  paired = FALSE,
  nnorm = 3
)
}
\arguments{
\item{datafolder}{is the folder where the data is stored, downloaded using the getTCGA_RNAseqData().}

\item{geneSymbol}{is a vector of gene names to be analyzed.}

\item{projects}{is TCGA project.}

\item{geneType}{refers to the fil_col in the filterGeneTypeExpr() function}

\item{dataType}{is one of 'tpm', 'fpkm', or 'count'.}

\item{pattern}{is a pattern regular expression used to match the data files in the datafolder.}

\item{paired}{specifies whether to only retrieve data from paired samples.}

\item{nnorm}{indicates the minimum number of normal samples that must be included.}
}
\value{
a data.frame
}
\description{
getGeneExpData.pancancer
}
