% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFileProcessing.R
\name{getGeneInfoFromeGTF}
\alias{getGeneInfoFromeGTF}
\title{getGeneInfoFromeGTF}
\usage{
getGeneInfoFromeGTF(gtf)
}
\arguments{
\item{gtf}{gtf gtf Gene annotation file in gtf format. https://www.gencodegenes.org/}
}
\value{
a data.frame
}
\description{
getGeneInfoFromeGTF
}
