% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFileProcessing.R
\name{getGeneLenFromeGTF}
\alias{getGeneLenFromeGTF}
\title{getGeneLenFromeGTF}
\usage{
getGeneLenFromeGTF(gtf)
}
\arguments{
\item{gtf}{gtf Gene annotation file in gtf format. https://www.gencodegenes.org/}
}
\value{
a data.fram
}
\description{
getGeneLenFromeGTF
}
\seealso{
getGeneInfoFromeGTF,getGeneBaseInfo
}
