% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{getGeneTypeInfoFromeGTF}
\alias{getGeneTypeInfoFromeGTF}
\title{getGeneTypeInfoFromeGTF}
\usage{
getGeneTypeInfoFromeGTF(gtf)
}
\arguments{
\item{gtf}{The gtf parameter represents the path to a GTF (Gene Transfer Format) format file.}
}
\value{
data.frame
}
\description{
getGeneTypeInfoFromeGTF
}
