% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{getInfiltDataOfTCGAsample}
\alias{getInfiltDataOfTCGAsample}
\title{getInfiltDataOfTCGAsample}
\usage{
getInfiltDataOfTCGAsample(
  expr,
  idtype = "barcode",
  datatype = "tumor",
  TIMER2,
  method
)
}
\arguments{
\item{expr}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{idtype}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{datatype}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TIMER2}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
getInfiltDataOfTCGAsample
}
