% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_dataDownload.R
\name{getTCGA_miRNA_IsoformEQ}
\alias{getTCGA_miRNA_IsoformEQ}
\title{getTCGA_miRNA_IsoformEQ}
\usage{
getTCGA_miRNA_IsoformEQ(project, save = FALSE, folder = ".")
}
\arguments{
\item{project}{A list of valid project (see list with TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{save}{TRUE or FALSE}

\item{folder}{The path specifies the location of the folder where the data is saved when the 'save' parameter is set to TRUE.}
}
\value{
data.frame
}
\description{
getTCGA_miRNA_IsoformEQ
}
