% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalysis.R
\name{getTCGAgeneExpDat}
\alias{getTCGAgeneExpDat}
\title{getTCGAgeneExpDat}
\usage{
getTCGAgeneExpDat(
  datafolder,
  geneSymbol,
  projects = "All",
  geneType = "protein_coding",
  dataType = "tpm",
  pattern = "STARdata.Rdata$"
)
}
\arguments{
\item{datafolder}{is the folder where the data is stored, downloaded using the getTCGA_RNAseqData().}

\item{geneSymbol}{is a vector of gene names to be analyzed.}

\item{projects}{is TCGA project.}

\item{geneType}{refers to the fil_col in the filterGeneTypeExpr() function}

\item{dataType}{is one of 'tpm', 'fpkm', or 'count'.}

\item{pattern}{is a pattern regular expression used to match the data files in the datafolder.}
}
\value{
data.frame
}
\description{
getTCGAgeneExpDat
}
