% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalysis.R
\name{ggplotGenePancancerExp}
\alias{ggplotGenePancancerExp}
\title{ggplotGenePancancerExp}
\usage{
ggplotGenePancancerExp(data, gene, save = FALSE, folder = ".", paired = FALSE)
}
\arguments{
\item{data}{is obtained from the getGeneExpData.fancancer function, representing the gene expression data.}

\item{gene}{is a string representing a specific gene.}

\item{save}{TRUE or FALSE}

\item{folder}{The path specifies the location of the folder where the data is saved when the 'save' parameter is set to TRUE.}

\item{paired}{indicates whether the data consists of paired samples.}
}
\value{
ggolot object.
}
\description{
ggplotGenePancancerExp
}
