% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{mergeSurExp}
\alias{mergeSurExp}
\title{mergeSurExp}
\usage{
mergeSurExp(
  expr,
  survival,
  Timeunit = 1,
  survivalFrome = NULL,
  TCGA = FALSE,
  TCGAfrome = "MedBioInfoCloud",
  feature = NULL,
  save = FALSE,
  folder = "."
)
}
\arguments{
\item{expr}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{survival}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{Timeunit}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{survivalFrome}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGAfrome}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{feature}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
a data.frame
}
\description{
mergeSurExp
}
