% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{Met450prepare}
\alias{Met450prepare}
\title{Met450prepare}
\usage{
Met450prepare(query)
}
\arguments{
\item{query}{The GDCquery function requests the return result of the methylation data.}
}
\value{
data.frame
}
\description{
Met450prepare
}
