% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{MultivariateCOX}
\alias{MultivariateCOX}
\title{MultivariateCOX}
\usage{
MultivariateCOX(
  data,
  dataFrom = "mergeSurExp",
  feature = "all",
  train_prop = 0.8,
  cutoff = 0.05,
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{dataFrom}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{feature}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{train_prop}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cutoff}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
MultivariateCOX
}
