% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{MultivariateCOX.verify}
\alias{MultivariateCOX.verify}
\title{MultivariateCOX.verify}
\usage{
MultivariateCOX.verify(dataset, MulCox.sigFactors, coef)
}
\arguments{
\item{dataset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{MulCox.sigFactors}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{coef}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\description{
MultivariateCOX.verify
}
