% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqVisual.R, R/visual_point.R,
%   R/visualization.R
\name{plotDEGvolcanoFig}
\alias{plotDEGvolcanoFig}
\title{plotDEGvolcanoFig}
\usage{
plotDEGvolcanoFig(
  data,
  x,
  y,
  cut_pvalue,
  cutFC,
  title,
  group,
  label,
  colour = c("#DC143C", "#00008B", "#808080")
)

plotDEGvolcanoFig(
  data,
  x,
  y,
  cut_pvalue,
  cutFC,
  title,
  group,
  label,
  colour = c("#DC143C", "#00008B", "#808080")
)

plotDEGvolcanoFig(
  data,
  x,
  y,
  cut_pvalue,
  cutFC,
  title,
  group,
  label,
  colour = c("#DC143C", "#00008B", "#808080")
)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{x}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{y}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cut_pvalue}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cutFC}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{title}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{group}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{label}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{colour}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.

For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.

For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
plotDEGvolcanoFig

plotDEGvolcanoFig

plotDEGvolcanoFig
}
