% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{read.gmt.to.getGmt}
\alias{read.gmt.to.getGmt}
\title{read.gmt.to.getGmt}
\usage{
read.gmt.to.getGmt(genesetdf)
}
\arguments{
\item{genesetdf}{The result of the read.gmt() function}
}
\value{
GeneSetCollection object.
}
\description{
read.gmt.to.getGmt
}
