% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{read_copy_number_variation}
\alias{read_copy_number_variation}
\title{read_copy_number_variation}
\usage{
read_copy_number_variation(files, cases)
}
\arguments{
\item{files}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cases}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\description{
read_copy_number_variation
}
