% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{read_gene_level_copy_number}
\alias{read_gene_level_copy_number}
\title{read_gene_level_copy_number}
\usage{
read_gene_level_copy_number(files, cases, summarizedExperiment = FALSE)
}
\arguments{
\item{files}{file path.}

\item{cases}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{summarizedExperiment}{}
}
\value{
data.frame
}
\description{
read_gene_level_copy_number
}
