% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{readGISTIC}
\alias{readGISTIC}
\title{readGISTIC}
\usage{
readGISTIC(files, cases)
}
\arguments{
\item{files}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cases}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
readGISTIC
}
