% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{readMet450data}
\alias{readMet450data}
\title{readMet450data}
\usage{
readMet450data(project = NULL, dataPath = "GDCdata", json = NULL)
}
\arguments{
\item{project}{A list of valid project (see list with TCGAbiolinks:::getGDCprojects()$project_id)]}

\item{dataPath}{"GDCdata"}

\item{json}{NULL}
}
\value{
data.frame
}
\description{
readMet450data
}
