% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{RNAseqDataConversion}
\alias{RNAseqDataConversion}
\title{RNAseqDataConversion}
\usage{
RNAseqDataConversion(data, type, species = "homo", gtf = NULL)
}
\arguments{
\item{data}{RNAseq expression data, either a matrix or a data frame}

\item{type}{must be one of "Counts2TPM", "Counts2FPKM", or "FPKM2TPM"}

\item{species}{must be one of "homo", "mus", or NULL.}

\item{gtf}{The gtf parameter represents the path to a GTF (Gene Transfer Format) format file.}
}
\value{
either a matrix or a data frame
}
\description{
RNAseqDataConversion
}
