% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{RNAseqDataTrans}
\alias{RNAseqDataTrans}
\title{RNAseqDataTrans}
\usage{
RNAseqDataTrans(data, tfun, species, gtype)
}
\arguments{
\item{data}{a data.frame or matrix for raw count or FPKM of RNAseq}

\item{tfun}{tfun only support 'Counts2FPKM','FPKM2TPM' or 'Counts2TPM'}

\item{species}{'hsa' or 'mus'}

\item{gtype}{'Ensemble' or 'Symbol'}
}
\description{
RNAseqDataTrans
}
