% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseAnalyses.R
\name{scoringSys}
\alias{scoringSys}
\title{scoringSys}
\usage{
scoringSys(
  expr,
  geneset,
  TCGA = FALSE,
  method = "ssgsea",
  filtergenenum = 10,
  study.type = "bulk_RNAseq",
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{expr}{TPM matrix derived from scRNAseq, or normalized gene expression matrix from bulk samples; rows are genes, cols are cells or samples.}

\item{geneset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{study.type}{A string value to indicate the study type for calculation. Allowed values contain c('scRNAseq', 'bulk_RNAseq').}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
data.frame
}
\description{
scoringSys
}
