% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAassist.R
\name{splitTCGAmatrix}
\alias{splitTCGAmatrix}
\title{splitTCGARNAseqdata}
\usage{
splitTCGAmatrix(data, sample = "Tumor")
}
\arguments{
\item{data}{data.frame or matrix, the columns should be the barcode of the TCGA patient sample}

\item{sample}{"Tumor" or "Normal"}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
splitTCGARNAseqdata
}
