% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAanalysis.R
\name{TCGA_Clin_Analysis}
\alias{TCGA_Clin_Analysis}
\title{TCGA_Clin_Analysis}
\usage{
TCGA_Clin_Analysis(
  data,
  clin_feature,
  feature,
  title = "",
  width = 5,
  height = 5,
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{clin_feature}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{feature}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{title}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{width}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{height}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
TCGA_Clin_Analysis
}
