% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{tidyGene.fromeReactome}
\alias{tidyGene.fromeReactome}
\title{tidyGene.fromeReactome}
\usage{
tidyGene.fromeReactome(
  filepath,
  fun = "stat",
  Source = "Reactome",
  addTotal = TRUE,
  termName = NULL,
  save = TRUE,
  folder = ".",
  filename = "geneset"
)
}
\arguments{
\item{filepath}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{fun}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{Source}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{termName}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filename}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
data.frame
}
\description{
tidyGene.fromeReactome
}
