% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrognosticModel.R
\name{UnivariateCOX}
\alias{UnivariateCOX}
\title{UnivariateCOX}
\usage{
UnivariateCOX(data, variable)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{variable}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
data.frame
}
\description{
UnivariateCOX
}
