% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA.blockwiseModules}
\alias{WGCNA.blockwiseModules}
\title{WGCNA.blockwiseModules}
\usage{
WGCNA.blockwiseModules(expr, recal = FALSE, folder = ".")
}
\arguments{
\item{expr}{TPM matrix.}

\item{recal}{TURE or FALSE.}

\item{folder}{Folder path}
}
\value{
list
}
\description{
WGCNA.blockwiseModules
}
