% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA.hubgeneBasedgeneCor}
\alias{WGCNA.hubgeneBasedgeneCor}
\title{WGCNA.hubgeneBasedgeneCor}
\usage{
WGCNA.hubgeneBasedgeneCor(
  expr,
  gene,
  TCGA = FALSE,
  filter = "protein_coding",
  cor.method = "pearson",
  cor.coef = 0.3,
  cor.p.value = 0.05,
  phenotype = NULL,
  geneset = NULL,
  method = "ssgsea",
  study.type = "bulk_RNAseq",
  heatmapH = 6,
  heatmapW = 14,
  mtc = 0.7,
  MM = 0.6,
  GS = 0.6,
  save = TRUE,
  folder = "."
)
}
\arguments{
\item{expr}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{gene}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filter}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.coef}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{cor.p.value}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{phenotype}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{geneset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{study.type}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapH}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapW}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{MM}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{GS}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{moduleTraitCor}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
WGCNA.hubgeneBasedgeneCor
}
