% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA.ModulesPhenotype}
\alias{WGCNA.ModulesPhenotype}
\title{WGCNA.ModulesPhenotype}
\usage{
WGCNA.ModulesPhenotype(
  expr,
  phenotype,
  TCGA = FALSE,
  filter = "protein_coding",
  recal = FALSE,
  outputNet = TRUE,
  colors = blueWhiteRed(50),
  heatmapH = NULL,
  heatmapW = NULL,
  mtc = 0.7,
  MM = 0.6,
  GS = 0.6,
  folder = "."
)
}
\arguments{
\item{expr}{TPM matrix.}

\item{phenotype}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filter}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{recal}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{colors}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapH}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapW}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{MM}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{GS}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{moduleTraitCor}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
finish
}
\description{
WGCNA.ModulesPhenotype
}
