% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA.R
\name{WGCNA.ModulesScoresys}
\alias{WGCNA.ModulesScoresys}
\title{WGCNA.ModulesScoresys}
\usage{
WGCNA.ModulesScoresys(
  expr,
  TCGA = FALSE,
  geneset,
  method = "ssgsea",
  study.type = "bulk_RNAseq",
  heatmapH = 6,
  heatmapW = 14,
  mtc = 0.7,
  MM = 0.6,
  GS = 0.6,
  folder = "."
)
}
\arguments{
\item{expr}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{TCGA}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{geneset}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{method}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{study.type}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapH}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{heatmapW}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{MM}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{GS}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{moduleTraitCor}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.
}
\description{
WGCNA.ModulesScoresys
}
