% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{arrayDataDEA_limma}
\alias{arrayDataDEA_limma}
\title{arrayDataDEA_limma}
\usage{
arrayDataDEA_limma(data, group, comparison)
}
\arguments{
\item{data}{matrix}

\item{group}{a vector}

\item{comparison}{character}
}
\value{
data.frame
}
\description{
arrayDataDEA_limma
}
