% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichVisual.R
\name{common_bar}
\alias{common_bar}
\title{common_bar}
\usage{
common_bar(data, palette, axisTitle.x, axisTitle.y, title)
}
\arguments{
\item{data}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{palette}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{axisTitle.x}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{axisTitle.y}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{title}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
common_bar
}
