% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{geneset2gmt}
\alias{geneset2gmt}
\title{geneset2gmt}
\usage{
geneset2gmt(
  geneset,
  genesetname,
  description = NA,
  return = "data.frame",
  folder = ".",
  filename
)
}
\arguments{
\item{geneset}{A gene vector or a txt file path. Note that the contents of the txt text file should be one gene in a row.}

\item{genesetname}{A string meaning the name of the gene set.}

\item{description}{A string used to introduce the gene set, default is NA.}

\item{return}{Return a value of "data frame" or "GeneSetCollection", said what kind of value, the equivalent of respectively with clusterProfiler: : read. GMT () function and GSEABase: : getGmt () function to read in the result of the GMT file, You can return the result as required.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filename}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
"data frame" or "GeneSetCollection" object.
}
\description{
geneset2gmt
}
