% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFileProcessing.R
\name{getGeneBaseInfo}
\alias{getGeneBaseInfo}
\title{getGeneBaseInfo}
\usage{
getGeneBaseInfo(gtf)
}
\arguments{
\item{gtf}{gtf gtf Gene annotation file in gtf format. https://www.gencodegenes.org/}
}
\value{
a data.frame
}
\description{
getGeneBaseInfo
}
