% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFileProcessing.R, R/baseFun.R
\name{getGeneLenFromeGTF}
\alias{getGeneLenFromeGTF}
\title{getGeneLenFromeGTF}
\usage{
getGeneLenFromeGTF(gtf)

getGeneLenFromeGTF(gtf)
}
\arguments{
\item{gtf}{The gtf parameter represents the path to a GTF (Gene Transfer Format) format file.}
}
\value{
a data.fram

data.frame
}
\description{
getGeneLenFromeGTF

getGeneLenFromeGTF
}
\seealso{
getGeneInfoFromeGTF,getGeneBaseInfo
}
