% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichVisual.R
\name{GO_KEGG.enrichVisual_barplot}
\alias{GO_KEGG.enrichVisual_barplot}
\title{Title}
\usage{
GO_KEGG.enrichVisual_barplot(
  obj,
  showCategory = 6,
  palette = "RdPu",
  axisTitle.x = "Number of Gene",
  axisTitle.y = "Term name",
  title = "Enrichment barplot",
  save = FALSE,
  folder = "./",
  fileName = "EnrichBar",
  height = 6,
  width = 10
)
}
\arguments{
\item{showCategory}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{palette}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{axisTitle.x}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{axisTitle.y}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{title}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{save}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{fileName}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{height}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{width}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{enrichResult}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}
}
\value{
list
}
\description{
Title
}
