% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFun.R
\name{outputGmtFile}
\alias{outputGmtFile}
\title{outputGmtFile}
\usage{
outputGmtFile(input, description = NA, folder = ".", filename)
}
\arguments{
\item{input}{list or data.frame}

\item{description}{NA}

\item{folder}{For more learning materials, please refer to https://github.com/BioInfoCloud/MedBioInfoCloud.}

\item{filename}{filepath}
}
\description{
outputGmtFile
}
