% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalyses.R
\name{selectDEG}
\alias{selectDEG}
\title{Screening for differentially expressed genes}
\usage{
selectDEG(DEAR, cutFC = 1, pMethod = "FDR", cutP = 0.05)
}
\arguments{
\item{DEAR}{a data.frame. The results of differential expression analysis,
must contain the column specified by `pMethod` (e.g., "FDR", "pAdj", "pValue")
and a "log2FC" column.}

\item{cutFC}{The threshold absolute value for log2 fold change.
The default is 1.}

\item{pMethod}{Metrics for statistical testing: "FDR", "pAdj" or "pValue".}

\item{cutP}{Significance screening threshold. Default value: 0.05}
}
\value{
data.frame with an additional "direction" column indicating
  Up/Down/Ns (Not significant)
}
\description{
Screening for differentially expressed genes
}
